% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_db_comparison.R
\name{compare_dataframes}
\alias{compare_dataframes}
\title{Compare two data.frames: throws a warning if not identical and returns the not identical rows}
\usage{
compare_dataframes(df1, df2, allowed_missings_df2 = NULL)
}
\arguments{
\item{df1}{data.frame 1}

\item{df2}{data.frame 2}

\item{allowed_missings_df2}{tables which are allowed to be missing in df2}
}
\description{
Compare two data.frames: throws a warning if not identical and returns the not identical rows
}
