% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_db_comparison.R
\name{compare_dev_prod}
\alias{compare_dev_prod}
\title{Compares dev and prod data base on clickhouse}
\usage{
compare_dev_prod(ch2_data)
}
\arguments{
\item{ch2_data}{data.frame of tables or column types available on the clickhouse 2 db}
}
\description{
Compares dev and prod data base on clickhouse
}
