% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_db_comparison.R
\name{send_query_local}
\alias{send_query_local}
\title{Send query wrapper to a clickhouse instance available on localhost}
\usage{
send_query_local(query)
}
\arguments{
\item{query}{query as string or file}
}
\description{
Send query wrapper to a clickhouse instance available on localhost
}
