% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_db_comparison.R
\name{send_query_prod}
\alias{send_query_prod}
\title{Send query wrapper to clickhouse}
\usage{
send_query_prod(query)
}
\arguments{
\item{query}{query as string or file}
}
\description{
Send query wrapper to clickhouse
}
