% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_sources.R
\name{data_sources_grid}
\alias{data_sources_grid}
\alias{data_sources_grid_sim}
\alias{data_sources_stations}
\alias{s3_data_sources}
\alias{cleanup_data_sources}
\alias{single_source}
\title{Data sources to query from clickhouse}
\usage{
data_sources_grid()

data_sources_grid_sim()

data_sources_stations()

s3_data_sources(year = format(Sys.Date(), "\%Y"))

cleanup_data_sources()

single_source(
  name,
  get_data = fairqDbtools::send_query,
  cleanup = FALSE,
  optimize = FALSE,
  skip_extract = FALSE,
  ...
)
}
\arguments{
\item{year}{year part of the current date, so that old years are preserved
but updates within the year overwrite the old data}

\item{name}{(character) a unique name used as class and to identify target
table on database.}

\item{get_data}{(function)}

\item{cleanup}{is there a cleanup step, i.e. a query with cleanup-\link{name}.sql to execute beforehand?}

\item{optimize}{is there a optimize step, i.e. a query with optimize-\link{name}.sql to execute after?}

\item{skip_extract}{should the extract step be skipped?}

\item{...}{arguments used within methods and available for queries}
}
\description{
List data sources. This list defines which data source to extract from clickhouse.

List data sources. This list defines which data source to extract from clickhouse.

List data sources. This list defines which data source to extract from clickhouse.

List data sources. This list defines which data source to extract from clickhouse.
}
